+++ R/group-by.R
@@ -55,13 +55,15 @@
 #'
 #' ## Legacy behavior
 #'
+#' `r lifecycle::badge("deprecated")`
+#'
 #' Prior to dplyr 1.1.0, character vector grouping columns were ordered in the
-#' system locale. If you need to temporarily revert to this behavior, you can
-#' set the global option `dplyr.legacy_locale` to `TRUE`, but this should be
-#' used sparingly and you should expect this option to be removed in a future
-#' version of dplyr. It is better to update existing code to explicitly call
-#' `arrange(.locale = )` instead. Note that setting `dplyr.legacy_locale` will
-#' also force calls to [arrange()] to use the system locale.
+#' system locale. Setting the global option `dplyr.legacy_locale` to `TRUE`
+#' retains this legacy behavior, but this has been deprecated. Update existing
+#' code to explicitly call `arrange(.locale = )` instead. Run
+#' `Sys.getlocale("LC_COLLATE")` to determine your system locale, and compare
+#' that against the list in [stringi::stri_locale_list()] to find an appropriate
+#' value for `.locale`, i.e. for American English, `"en_US"`.
 #'
 #' @export
 #' @examples