+++ tests/testthat/test-arrange.R
@@ -326,7 +326,7 @@ test_that("arrange keeps zero length groups", {
 
 # legacy --------------------------------------------------------------
 
-test_that("legacy - using the global option `dplyr.legacy_locale` forces the system locale", {
+test_that("legacy - using the deprecated global option `dplyr.legacy_locale` forces the system locale", {
   skip_if_not(has_collate_locale("en_US"), message = "Can't use 'en_US' locale")
 
   local_options(dplyr.legacy_locale = TRUE)
@@ -334,12 +334,18 @@ test_that("legacy - using the global option `dplyr.legacy_locale` forces the sys
 
   df <- tibble(x = c("a", "A", "Z", "b"))
 
-  expect_identical(arrange(df, x)$x, c("a", "A", "b", "Z"))
+  # Capture `dplyr.legacy_locale` deprecation warning
+  expect_snapshot({
+    out <- arrange(df, x)$x
+  })
+
+  expect_identical(out, c("a", "A", "b", "Z"))
 })
 
 test_that("legacy - usage of `.locale` overrides `dplyr.legacy_locale`", {
   skip_if_not_installed("stringi", "1.5.3")
 
+  local_options(lifecycle_verbosity = "quiet")
   local_options(dplyr.legacy_locale = TRUE)
 
   # Danish `o` with `/` through it sorts after `z` in Danish locale
@@ -355,6 +361,7 @@ test_that("legacy - usage of `.locale` overrides `dplyr.legacy_locale`", {
 })
 
 test_that("legacy - empty arrange() returns input", {
+  local_options(lifecycle_verbosity = "quiet")
   local_options(dplyr.legacy_locale = TRUE)
 
   df <- tibble(x = 1:10, y = 1:10)
@@ -368,6 +375,7 @@ test_that("legacy - empty arrange() returns input", {
 })
 
 test_that("legacy - can sort empty data frame", {
+  local_options(lifecycle_verbosity = "quiet")
   local_options(dplyr.legacy_locale = TRUE)
 
   df <- tibble(a = numeric(0))
@@ -375,6 +383,7 @@ test_that("legacy - can sort empty data frame", {
 })
 
 test_that("legacy - local arrange sorts missing values to end", {
+  local_options(lifecycle_verbosity = "quiet")
   local_options(dplyr.legacy_locale = TRUE)
 
   df <- data.frame(x = c(2, 1, NA))
@@ -384,6 +393,7 @@ test_that("legacy - local arrange sorts missing values to end", {
 })
 
 test_that("legacy - arrange handles list columns (#282)", {
+  local_options(lifecycle_verbosity = "quiet")
   local_options(dplyr.legacy_locale = TRUE)
 
   # no intrinsic ordering
@@ -395,6 +405,7 @@ test_that("legacy - arrange handles list columns (#282)", {
 })
 
 test_that("legacy - arrange handles raw columns (#1803)", {
+  local_options(lifecycle_verbosity = "quiet")
   local_options(dplyr.legacy_locale = TRUE)
 
   df <- tibble(x = 1:3, y = as.raw(3:1))
@@ -402,6 +413,7 @@ test_that("legacy - arrange handles raw columns (#1803)", {
 })
 
 test_that("legacy - arrange handles matrix columns", {
+  local_options(lifecycle_verbosity = "quiet")
   local_options(dplyr.legacy_locale = TRUE)
 
   df <- tibble(x = 1:3, y = matrix(6:1, ncol = 2))
@@ -409,6 +421,7 @@ test_that("legacy - arrange handles matrix columns", {
 })
 
 test_that("legacy - arrange handles data.frame columns (#3153)", {
+  local_options(lifecycle_verbosity = "quiet")
   local_options(dplyr.legacy_locale = TRUE)
 
   df <- tibble(x = 1:3, y = data.frame(z = 3:1))
@@ -416,6 +429,7 @@ test_that("legacy - arrange handles data.frame columns (#3153)", {
 })
 
 test_that("legacy - arrange handles complex columns", {
+  local_options(lifecycle_verbosity = "quiet")
   local_options(dplyr.legacy_locale = TRUE)
 
   df <- tibble(x = 1:3, y = 3:1 + 2i)
@@ -423,6 +437,7 @@ test_that("legacy - arrange handles complex columns", {
 })
 
 test_that("legacy - arrange handles S4 classes (#1105)", {
+  local_options(lifecycle_verbosity = "quiet")
   local_options(dplyr.legacy_locale = TRUE)
 
   TestS4 <- suppressWarnings(setClass("TestS4", contains = "integer"))
@@ -436,6 +451,7 @@ test_that("legacy - arrange handles S4 classes (#1105)", {
 })
 
 test_that("legacy - `arrange()` works with `numeric_version` (#6680)", {
+  local_options(lifecycle_verbosity = "quiet")
   local_options(dplyr.legacy_locale = TRUE)
 
   x <- numeric_version(c("1.11", "1.2.3", "1.2.2"))
@@ -447,6 +463,7 @@ test_that("legacy - `arrange()` works with `numeric_version` (#6680)", {
 })
 
 test_that("legacy - arrange works with two columns when the first has a data frame proxy (#6268)", {
+  local_options(lifecycle_verbosity = "quiet")
   local_options(dplyr.legacy_locale = TRUE)
 
   # `id1` has a data frame proxy for `vec_proxy_order()`
@@ -461,6 +478,7 @@ test_that("legacy - arrange works with two columns when the first has a data fra
 })
 
 test_that("legacy - arrange() supports across() and pick() (#4679)", {
+  local_options(lifecycle_verbosity = "quiet")
   local_options(dplyr.legacy_locale = TRUE)
 
   df <- tibble(x = c(1, 3, 2, 1), y = c(4, 3, 2, 1))
@@ -484,6 +502,7 @@ test_that("legacy - arrange() supports across() and pick() (#4679)", {
 })
 
 test_that("legacy - arrange() works with across() and pick() cols that return multiple columns (#6490)", {
+  local_options(lifecycle_verbosity = "quiet")
   local_options(dplyr.legacy_locale = TRUE)
 
   df <- tibble(
@@ -508,6 +527,7 @@ test_that("legacy - arrange() works with across() and pick() cols that return mu
 })
 
 test_that("legacy - arrange sorts missings in df-cols correctly", {
+  local_options(lifecycle_verbosity = "quiet")
   local_options(dplyr.legacy_locale = TRUE)
 
   col <- tibble(a = c(1, 1, 1), b = c(3, NA, 1))
@@ -518,6 +538,7 @@ test_that("legacy - arrange sorts missings in df-cols correctly", {
 })
 
 test_that("legacy - arrange with duplicates in a df-col uses a stable sort", {
+  local_options(lifecycle_verbosity = "quiet")
   local_options(dplyr.legacy_locale = TRUE)
 
   col <- tibble(a = c(1, 1, 1, 1, 1), b = c(3, NA, 2, 3, NA))
@@ -528,6 +549,7 @@ test_that("legacy - arrange with duplicates in a df-col uses a stable sort", {
 })
 
 test_that("legacy - arrange with doubly nested df-col doesn't infloop", {
+  local_options(lifecycle_verbosity = "quiet")
   local_options(dplyr.legacy_locale = TRUE)
 
   one <- tibble(a = c(1, 1, 1, 1, 1), b = c(1, 1, 2, 2, 2))